% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWQMS_chars.R
\docType{data}
\name{AWQMS_chars}
\alias{AWQMS_chars}
\title{Table of AWQMS characters values.}
\format{An object of class \code{data.frame} with 6591 rows and 3 columns.}
\usage{
data(AWQMS_chars)
}
\description{
Table of AWQMS characters values.
}
\examples{
data(AWQMS_chars)
}
\keyword{AWQMS}
\keyword{datasets}
