% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wqdb.R
\name{create_wqdb}
\alias{create_wqdb}
\title{Create a wqdb formatted SQLite database.}
\usage{
create_wqdb(db)
}
\arguments{
\item{db}{The path and file name to the new SQLite database to be created.}
}
\value{
None
}
\description{
Create a SQLite database with table fields identical to Oregon DEQ's AWQMS and Stations database.
If the database already exists this function will check if the tables exist and create them if not.
Tables created include:
'stations':  Table of station information
'characteristics' Table of AWQMS charateristics. Same as the table returned from wqdblite::AWQMS_chars()
'awqms' Table of AWQMS data
'other' Table of non AWQMS data
'continuous' Table of continous data
'vw_discrete' view of awqms and other outer joined with Stations
'vw_continuous' view of continuous outer joined with Stations
}
\keyword{database,}
\keyword{sqlite}
