% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_stations.R
\name{import_stations}
\alias{import_stations}
\title{Take a vector of monitoring location station IDs (MLocIDs) and query station data from DEQ's Stations database. This station data is then loaded in the Stations table in the specified SQLite database. Note - trying to load in duplicate values will cause an error.}
\usage{
import_stations(stations, sqlite_db)
}
\arguments{
\item{stations}{Dataframe of monitoring location station information to import into the SQLite database. The table format should be the same as Oregon DEQ Stations database.}

\item{sqlite_db}{The path and file name to the SQLite database where the stations data will be imported into.}
}
\description{
Take a vector of monitoring location station IDs (MLocIDs) and query station data from DEQ's Stations database. This station data is then loaded in the Stations table in the specified SQLite database. Note - trying to load in duplicate values will cause an error.
}
\keyword{stations}
