% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wqdb.R
\name{read_wqdb}
\alias{read_wqdb}
\title{Retrive data from a water quality SQLite database.}
\usage{
read_wqdb(sqlite_db, startdate = "1949-09-15", enddate = NULL,
  station = NULL, project = NULL, char = NULL, stat_base = NULL,
  media = NULL, org = NULL, HUC8 = NULL, HUC8_Name = NULL,
  HUC10 = NULL, HUC12 = NULL, HUC12_Name = NULL)
}
\arguments{
\item{sqlite_db}{The path and file name to the SQLite database.}

\item{startdate}{Required parameter setting the startdate of the data being fetched. Format 'yyyy-mm-dd'}

\item{enddate}{Optional parameter setting the enddate of the data being fetched. Format 'yyyy-mm-dd'}

\item{station}{Optional vector of stations to be fetched}

\item{project}{Optional vector of projects to be fetched}

\item{char}{Optional vector of characteristics to be fetched}

\item{stat_base}{Optional vector of Result Stattistical Bases to be fetched ex. Maximum}

\item{media}{Optional vector of sample media to be fetched}

\item{org}{optional vector of Organizations to be fetched}

\item{HUC8}{Optional vector of HUC8 codes to be fetched}

\item{HUC8_Name}{Optional vector of HUC8 names to be fetched}

\item{HUC10}{Optional vector of HUC10s to be fetched}

\item{HUC12}{Optional vector of HUC12s to be fetched}

\item{HUC12_Name}{Optional vector of HUC12 names to be fetched}

\item{crit_codes}{If true, include standard codes used in determining criteria}
}
\value{
Dataframe from the water quality database
}
\description{
Retrive data from a water quality SQLite database. Queries the vw_data_all view only. Function is a SQLite version of the AWQMS_Data function from the AWQMS_Data package.
}
