% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_wqdb.R
\name{write_wqdb}
\alias{write_wqdb}
\title{Write data into a wqdb formatted SQLite Database.}
\usage{
write_wqdb(db, df = NULL, table = "awqms")
}
\arguments{
\item{db}{The path and file name to the new SQLite database to be created.}

\item{df}{The dataframe to be written into the SQLite database. Default value is NULL.
The dataframe columns and datatypes are the same as a dataframe returned from AWQMSdata::AWQMS_data().}

\item{table}{The db table where df will be written. By default this is set to "awqms". Other options inlcude "stations",
"other", "continuous", or "characteristics".
"characteristics"
`chr_uid`: integer
`Char_Name`: character
`CASNumber`: character}
}
\value{
None
}
\description{
Create and/or write to a SQLite database with table fields identical to Oregon DEQ's AWQMS and Stations database.
If the database already exists this function will check if the tables exist and create them if not.
If a dataframe is passed the data will be written into the tables. Duplicate records are checked and not overwritten.
}
\keyword{database,}
\keyword{sqlite}
