% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wq_db.R
\name{create_wq_db}
\alias{create_wq_db}
\title{Create and format a water quality SQLite database.}
\usage{
create_wq_db(db)
}
\arguments{
\item{db}{The path and file name to the new SQLite database to be created.}
}
\value{
None
}
\description{
Create a SQLite database with table fields identical to Oregon DEQ's AWQMS and Stations database.
}
\keyword{database,}
\keyword{sqlite}
